pyproject.toml
[project]
name = "fastgen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2wsgi>=1.10.10",
    "aiofiles>=25.1.0",
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.13.2",
    "aiosignal>=1.4.0",
    "aiosqlite>=0.21.0",
    "alembic>=1.17.2",
    "altair>=6.0.0",
    "annotated-types>=0.7.0",
    "anyio>=4.12.0",
    "appdirs>=1.4.4",
    "asgiref>=3.11.0",
    "asttokens>=3.0.1",
    "attrs>=25.4.0",
    "av>=16.0.1",
    "beautifulsoup4>=4.14.3",
    "bentoml>=1.4.30",
    "black>=25.12.0",
    "blinker>=1.9.0",
    "cachetools>=6.2.2",
    "cattrs>=23.1.2",
    "certifi>=2025.11.12",
    "cffi>=2.0.0",
    "charset-normalizer>=3.4.4",
    "click>=8.3.1",
    "click-option-group>=0.5.9",
    "cloudpickle>=3.1.2",
    "colorama>=0.4.6",
    "comm>=0.2.3",
    "configargparse>=1.7",
    "dash>=3.3.0",
    "dash-core-components>=2.0.0",
    "dash-html-components>=2.0.0",
    "dash-table>=5.0.0",
    "decorator>=5.2.1",
    "deepmerge>=2.0",
    "deprecated>=1.3.1",
    "diffusers>=0.36.0",
    "distro>=1.9.0",
    "dnspython>=2.8.0",
    "dotenv>=0.9.9",
    "email-validator>=2.3.0",
    "executing>=2.2.1",
    "fastapi>=0.124.4",
    "fastapi-cli>=0.0.16",
    "fastjsonschema>=2.21.2",
    "filelock>=3.20.0",
    "flask>=3.1.2",
    "frozenlist>=1.8.0",
    "fs>=2.4.16",
    "fsspec>=2025.12.0",
    "gitdb>=4.0.12",
    "gitpython>=3.1.45",
    "google-genai>=1.55.0",
    "greenlet>=3.3.0",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "h11>=0.16.0",
    "h2>=4.3.0",
    "hpack>=4.1.0",
    "httpcore>=1.0.9",
    "httptools>=0.7.1",
    "httpx>=0.28.1",
    "httpx-ws>=0.8.2",
    "huggingface-hub>=0.36.0",
    "hyperframe>=6.1.0",
    "idna>=3.11",
    "importlib-metadata>=8.7.0",
    "inflection>=0.5.1",
    "ipython>=9.8.0",
    "ipywidgets>=8.1.8",
    "isort>=7.0.0",
    "itsdangerous>=2.2.0",
    "jedi>=0.19.2",
    "jinja2>=3.1.6",
    "jiter>=0.12.0",
    "jsonpatch>=1.33",
    "jsonpointer>=3.0.0",
    "jsonschema>=4.25.1",
    "jsonschema-specifications>=2025.9.1",
    "jupyter-core>=5.9.1",
    "jupyterlab-widgets>=3.0.16",
    "kantoku>=0.18.3",
    "langchain>=1.1.3",
    "langchain-core>=1.1.3",
    "langchain-openai>=1.1.2",
    "langchain-text-splitters>=1.0.0",
    "langsmith>=0.4.59",
    "loguru>=0.7.3",
    "mako>=1.3.10",
    "markdown-it-py>=4.0.0",
    "markupsafe>=3.0.3",
    "matplotlib-inline>=0.2.1",
    "mdurl>=0.1.2",
    "mpmath>=1.3.0",
    "multidict>=6.7.0",
    "mypy>=1.19.0",
    "mypy-extensions>=1.1.0",
    "narwhals>=2.13.0",
    "nbformat>=5.10.4",
    "nest-asyncio>=1.6.0",
    "networkx>=3.6.1",
    "numpy>=2.3.5",
    "nvidia-ml-py>=13.590.44",
    "open3d>=0.19.0",
    "openai>=2.11.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-instrumentation>=0.60b1",
    "opentelemetry-instrumentation-aiohttp-client>=0.60b1",
    "opentelemetry-instrumentation-asgi>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-semantic-conventions>=0.60b1",
    "opentelemetry-util-http>=0.60b1",
    "orjson>=3.11.5",
    "packaging>=25.0",
    "pandas>=2.3.3",
    "parso>=0.8.5",
    "pathspec>=0.12.1",
    "pillow>=12.0.0",
    "pip-requirements-parser>=32.0.1",
    "platformdirs>=4.5.1",
    "plotly>=6.5.0",
    "portalocker>=3.2.0",
    "prometheus-client>=0.23.1",
    "prompt-toolkit>=3.0.52",
    "propcache>=0.4.1",
    "protobuf>=6.33.2",
    "psutil>=7.1.3",
    "pure-eval>=0.2.3",
    "pyarrow>=22.0.0",
    "pycparser>=2.23",
    "pydantic>=2.12.5",
    "pydantic-core>=2.41.5",
    "pydantic-settings>=2.12.0",
    "pydeck>=0.9.1",
    "pygments>=2.19.2",
    "pyparsing>=3.2.5",
    "pypdf>=6.4.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "python-json-logger>=4.0.0",
    "python-multipart>=0.0.20",
    "pytz>=2025.2",
    "pyyaml>=6.0.3",
    "pyzmq>=27.1.0",
    "qdrant-client>=1.16.2",
    "questionary>=2.1.1",
    "referencing>=0.37.0",
    "regex>=2025.11.3",
    "requests>=2.32.5",
    "requests-toolbelt>=1.0.0",
    "retrying>=1.4.2",
    "rich>=14.2.0",
    "rich-toolkit>=0.17.0",
    "rpds-py>=0.30.0",
    "safetensors>=0.7.0",
    "schema>=0.7.8",
    "shellingham>=1.5.4",
    "simple-di>=0.1.5",
    "six>=1.17.0",
    "smmap>=5.0.2",
    "sniffio>=1.3.1",
    "soundfile>=0.13.1",
    "soupsieve>=2.8",
    "sqlalchemy>=2.0.45",
    "stack-data>=0.6.3",
    "starlette>=0.50.0",
    "streamlit>=1.52.1",
    "sympy>=1.14.0",
    "tenacity>=9.1.2",
    "tiktoken>=0.12.0",
    "tokenizers>=0.22.1",
    "toml>=0.10.2",
    "tomli-w>=1.2.0",
    "torch>=2.9.1",
    "tornado>=6.5.3",
    "tqdm>=4.67.1",
    "traitlets>=5.14.3",
    "transformers>=4.57.3",
    "typer>=0.20.0",
    "types-aiofiles>=25.1.0.20251011",
    "types-beautifulsoup4>=4.12.0.20250516",
    "types-html5lib>=1.1.11.20251117",
    "types-requests>=2.32.4.20250913",
    "typing-extensions>=4.15.0",
    "tzdata>=2025.2",
    "urllib3>=2.6.2",
    "uv>=0.9.17",
    "uvicorn>=0.38.0",
    "watchdog>=6.0.0",
    "watchfiles>=1.1.1",
    "wcwidth>=0.2.14",
    "websockets>=15.0.1",
    "werkzeug>=3.1.4",
    "widgetsnbextension>=4.0.15",
    "win32-setctime>=1.2.0",
    "wrapt>=1.17.3",
    "wsproto>=1.3.2",
    "yarl>=1.22.0",
    "zipp>=3.23.0",
]
